+++ src/lib/api/UserApi.js
@@ -54,3 +54,13 @@ export const userUpdatePassword = async (token, { password }) => {
     body: JSON.stringify({ password }),
   });
 };
+
+export const userLogout = async (token) => {
+  return fetch(`${import.meta.env.VITE_URL_API}/users/logout`, {
+    method: 'DELETE',
+    headers: {
+      Accept: 'application/json',
+      Authorization: token,
+    },
+  });
+};